+++ src/main/java/org/eclipse/keyple/core/util/json/CalendarJsonAdapter.java
@@ -39,7 +39,8 @@ public class CalendarJsonAdapter implements JsonSerializer<Calendar>, JsonDeseri
    * @since 2.2.0
    */
   public CalendarJsonAdapter() {
-    sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSXXX");
+    // We're using "Z" instead of "XXX" due to compatibility issues.
+    sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
     sdf.setTimeZone(TimeZone.getTimeZone("CET"));
   }
 
@@ -50,7 +51,15 @@ public class CalendarJsonAdapter implements JsonSerializer<Calendar>, JsonDeseri
    */
   @Override
   public JsonElement serialize(Calendar data, Type typeOfSrc, JsonSerializationContext context) {
-    return new JsonPrimitive(sdf.format(data.getTime()));
+    String formatted = sdf.format(data.getTime());
+
+    // We are manually inserting a colon in the timezone offset to have a ISO 8601 format and
+    // maintain compatibility with older versions of Java and Android that do not support the "XXX"
+    // timezone marker.
+    StringBuilder sb = new StringBuilder(formatted);
+    sb.insert(formatted.length() - 2, ':');
+    formatted = sb.toString();
+    return new JsonPrimitive(formatted);
   }
 
   /**